pyproject.toml
[project]
name = "manga-ocr-android"
version = "0.1.0"
description = "Model dev for manga-ocr-android"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-edge-torch-nightly (>=0.5.0.dev20250427,<0.6.0)",
    "click (>=8.1.8,<9.0.0)",
    "huggingface-hub (>=0.30.2,<0.31.0)",
    "onnx (>=1.17.0,<2.0.0)",
    "onnxruntime (>=1.21.1,<2.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "ultralytics (>=8.3.124,<9.0.0)",
    "onnxscript (>=0.2.5,<0.3.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "fugashi (>=1.4.0,<2.0.0)",
    "unidic-lite (>=1.0.8,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

packages = [{ include = "model-dev" }]
